requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
openai==1.53.0
pydantic==2.9.2
uvloop==0.20.0; platform_system != 'Windows'
